requirements.txt
numpy >= 1.23.5
opencv-python >= 4.7.0.72
matplotlib >= 3.7.1
seaborn >= 0.12.2
scipy >= 1.10.1
scikit-learn >= 1.2.2
tensorflow-gpu == 2.7.2
protobuf == 3.20.0
keras
pandas >= 2.0.1
Pillow >= 9.5.0
dvc-s3 >= 2.22.0
fsspec >= 2022.10.0
dvclive >= 2.8.1
nvgpu
lxml >= 4.9.2
python-collection >= 1.0.0